refactor: follow nextjs best practice
- use client and server component appropriately
